Added database migration functionality.
If the database schema changes as future versions of Sleet are
released, a migration SQL script can be created to migrate existing
databases to the new schema.

On startup, the version number in the database is compared with the
version number in the source code. If different, the necessary
migration scripts are executed. Each script updates the schema from one
version to the next (N+1) version.  For example, if the current schema
is version 1 and the newest schema is 3, then two migration scripts are
run (1 to 2, then 2 to 3).